requirements.txt
azure-storage-blob==12.9.0
azure-ai-formrecognizer==3.1.2
azure-identity==1.7.0
Werkzeug==2.3.7
Flask==2.1.0
fuzzywuzzy==0.18.0
openpyxl==3.0.9
pdf2image==1.16.0
pdfkit==1.0.0
polling2==0.5.0
PyPDF2==2.9.0
pymupdf==1.19.6
pytest==6.2.5
spacy==3.1.3
Pillow==9.2.0
gunicorn
tabula-py==2.6.0
typing_extensions<4.6.0
azure-servicebus==7.11.1
nltk==3.8.1
scikit-learn==1.1.3
joblib==1.3.2
requests==2.29.0
requests-oauthlib==1.3.1
urllib3==1.26.15
openai==0.27.8
langdetect==1.0.9
pyodbc==4.0.39
google-cloud-translate==3.15.3
img2pdf==0.5.1
progressbar==2.5
backoff==2.2.1
azure-keyvault==4.2.0
camelot-py==0.11.0
opencv-python==4.10.0.84
ghostscript==0.7
ABN==1.0.0
watchdog==4.0.2


